Introduce JediMark.
JediMark is a new blocktype with the jedi symbol on each side. When clicked on the top, it will auto-create a column of chisled stone 99 units high. Currently only available in creative mode.

Created in a 60 minute "How to make new blocks" lesson with Brian, Audrey, Flynn, and Max. For any new developers that want to experiment with new block types,
copy this class first and then modify it.