Jedi Mark: Welcome to the Dome!
New behavior for Jedi Mark:
(1) If clicked on a side, it builds an arc with radius=30 along the
same axis that you clicked (x or z axis).
(2) If you click the top, it builds a dome above you with radius=15.
(3) In both scenarios, the default material is chisled stone. You can
change this by putting the desired block-type in your hand when
clicking on the JediMark.