SithMark: Version 2
(1) Previously, the generated SithTower was hardcoded to always orient
south. Now it orients based on the side of the block that you hit (so
the door now will face whatever direction you clickd on the brick from).
(2) Various design changes to SithTower; twice as tall, with more
diagonal wall design.
(3) Now has secret feature: if you place two of these SithMarks in a
line (anywhere from 9-300 blocks away from each other), and the
left-most one is sitting on a stone block, then when you click the
side, you will get a monstrous SithCastle, which is 8 SithTowers
connected by walls.