Imperial Probe Droid improvements...
(1) introduced AI EnemyFactionDetector, which just detects a specific
faction in a given radius, and sets the attackTarget to the first one
it finds.
(2) modified AI Quickly Offended to not take offense if it is injured
by someone from the given friendly faction.
(3) modify ImperialProbeDroid with above new AI to have new behavior:
      - shoots snowballs at any jedi that approaches within 16 blocks
      - ignores any unaffiliated unless it damages them, in which case
it stands its ground and takes revenge until the enemy gets out of
range.
      - ignores empire, even if it is damaged by an empire player.